Update footer code structure
